requirements.txt
accounts==0.2.4
alembic==1.13.1
amqp==2.6.1
anyio==3.7.1
asgiref==3.7.2
attrs==23.2.0
bcrypt==4.1.2
billiard==3.6.4.0
cachetools==5.3.2
celery==4.4.7
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==7.1.2
cryptography==42.0.2
cssselect==1.2.0
cssutils==2.9.0
defusedxml==0.8.0rc2
Django==5.0.1
django-ckeditor==6.7.0
django-cors-headers==4.3.1
django-filter==23.5
django-jazzmin==2.6.0
django-js-asset==2.2.0
django-modeltranslation==0.18.11
django-phonenumber-field==7.3.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
dnspython==2.5.0
drf-yasg==1.21.7
ecdsa==0.18.0
email-validator==1.3.1
emails==0.5.15
exceptiongroup==1.2.0
fastapi==0.70.1
greenlet==3.0.3
gunicorn==20.1.0
h11==0.9.0
httptools==0.1.2
idna==3.6
inflect==5.6.2
inflection==0.5.1
Jinja2==2.11.3
kombu==4.6.11
lxml==5.1.0
Mako==1.3.2
MarkupSafe==2.1.5
more-itertools==10.2.0
oauthlib==3.2.2
packaging==23.2
passlib==1.7.4
phonenumbers==8.13.30
pillow==10.2.0
pluggy==0.13.1
premailer==3.10.0
psycopg2-binary==2.9.9
py==1.11.0
pyasn1==0.5.1
pycparser==2.21
pydantic==1.10.14
PyJWT==2.8.0
pytest==5.4.3
python-dateutil==2.8.2
python-jose==3.3.0
python-multipart==0.0.5
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
raven==6.10.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.4.0
social-auth-core==4.5.2
SQLAlchemy==1.4.51
sqlparse==0.4.4
starlette==0.16.0
tenacity==6.3.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.0
UserManager==0.5.3
uvicorn==0.11.8
uvloop==0.19.0
validators==0.22.0
vine==1.3.0
wcwidth==0.2.13
websockets==8.1
